requirements.txt
opencv-python==4.11.0.86
fpdf==1.7.2
numpy==2.2.3
pillow==11.1.0
streamlit==1.42.2
requests==2.32.3
